feat(content): widen map/sector coordinates and bridge Havenmere to 400x400
MapContent/MapDef.Width/Height and Sector.X/Y were byte, capping a map at
255 squares on a side. SEA_5 fixes every map at 400 squares
(WorldRules.MapSizeSquares), so ValidateWorldExtent could never pass:
400 does not fit in a byte. Widen all four to ushort, with checked casts
where Sector rows are seeded, and update the stale TerrainRows comment
that said a row index was "not a y yet" -- it is now, since the grid is
one row per square over the same 400-square world SectorRules already
addresses directly.

Havenmere's 20x20 authored grid is mechanically expanded 20x per axis
(every authored cell becomes a 20x20 block of the same terrain symbol,
every position and radius scaled to match) rather than redrawn by hand.
This is a bridge, not the final content: Phase 8 replaces it with a map
actually authored at 400x400. Seeding now writes 160,000 Sector rows for
this one map, against 400 before -- Init seeds every map, so this is the
real cost of the bridge, not a one-time charge.

The rescale widens current zone 1 to 56 squares, past the existing
28-square cap on authored current radius. That cap was never backed by a
SEA_5 figure (unlike the 40-square world-object cap, which is the spec's
storm size) -- it was just the widest zone the old content happened to
declare. Raising it to 56, documented as still unbacked by any
specification, was chosen over clamping the content because clamping
would silently misrepresent the bridged chart rather than admit the
placeholder it always was; Phase 8 sets the real ceiling when the maps
are authored by hand.

Extends the JSON content pipeline (scripts/lib/content-catalog.mjs) with
a ushort field kind, since MAP_FIELDS there is the source of truth for
generated field widths, not the generator script that reads it.

Coordinates hardcoded against the old 20x20 grid in SectorRulesTests and
ContentCatalogTests (sector lookups, terrain-row lengths, the port and
blocking-object placements) are updated to match the rescaled map; two of
these were already failing before this change and are fixed here rather
than left stale.

+++ server/spacetimedb/spacetimedb/Domain/ContentDefinitions.cs
@@ -30,8 +30,10 @@ public sealed record MapContent
     public required string Name { get; init; }
     public required string Biome { get; init; }
     public required byte MapRank { get; init; }
-    public required byte Width { get; init; }
-    public required byte Height { get; init; }
+    /// <summary>Squares across. SEA_5 §3.1 fixes this at 400 for every map.</summary>
+    public required ushort Width { get; init; }
+    /// <summary>Squares down.</summary>
+    public required ushort Height { get; init; }
     public required string PvpMode { get; init; }
     public required string MaterialId { get; init; }
     public required string PortName { get; init; }
@@ -43,11 +45,9 @@ public sealed record MapContent
     /// south, the same way the ruler and the world's own y axis run. There is no flip left
     /// between any of the three.
     ///
-    /// A row index is NOT a y yet. The authored grid is still 20 x 20 over a world that is
-    /// now 400 squares on a side, so one row spans twenty squares and
-    /// <see cref="SectorRules.TerrainAt"/> takes a scaled index, not a position. Feeding it
-    /// <c>floor(y)</c> throws once y passes 20. Task 1.6 expands the grid to one row per
-    /// square; only then does <c>TerrainRows[y][x]</c> read straight.
+    /// A row index is a y. The grid is one row per square over the 400-square world, so
+    /// <see cref="SectorRules.TerrainAt"/> takes a position directly: <c>TerrainRows[y][x]</c>
+    /// reads straight off <c>floor(y)</c> and <c>floor(x)</c>.
     /// </summary>
     public required IReadOnlyList<string> TerrainRows { get; init; }
     public required IReadOnlyList<WorldObjectContent> Objects { get; init; }
